requirements.txt
numpy>=1.26.1
Pillow>=10.1.0
robotpy>=2025.2.1.0